Move type Time to db
