+++ go.mod
@@ -7,5 +7,5 @@ require (
 	github.com/smartystreets/goconvey v1.7.2
 	github.com/stockparfait/errors v0.0.2
 	github.com/stockparfait/fetch v0.0.1
-	github.com/stockparfait/logging v0.0.1
+	github.com/stockparfait/logging v0.0.2
 )
